moving int to bigint.  Still won't compile.
git-svn-id: https://opendentalsoft.com:23793/svn/opendental/head@2989 098d0599-7f2b-0410-95a5-ad30ac85393a

+++ OpenDentBusiness/Data Interface/ClaimFormItems.cs
@@ -36,7 +36,7 @@ namespace OpenDentBusiness{
 		}
 
 		///<summary></summary>
-		public static int Insert(ClaimFormItem item){
+		public static long Insert(ClaimFormItem item) {
 			if(RemotingClient.RemotingRole==RemotingRole.ClientWeb) {
 				item.ClaimFormItemNum=Meth.GetInt(MethodBase.GetCurrentMethod(),item);
 				return item.ClaimFormItemNum;
