+++ OpenDentBusiness/Data Interface/Computers.cs
@@ -61,7 +61,7 @@ namespace OpenDentBusiness{
 		}
 
 		///<summary>ONLY use this if compname is not already present</summary>
-		public static int Insert(Computer comp){
+		public static long Insert(Computer comp) {
 			if(RemotingClient.RemotingRole==RemotingRole.ClientWeb) {
 				comp.ComputerNum=Meth.GetInt(MethodBase.GetCurrentMethod(),comp);
 				return comp.ComputerNum;
